add about, dokter, dokter-detail, home, and login logout function

+++ dokter/models.py
@@ -18,7 +18,7 @@ class Doctor(models.Model):
     Nama = models.CharField(max_length=100, blank=False)
     Jenkel = models.CharField(max_length=1, choices=Jenkel_LIST, blank=False)
     tgl_lahir = models.DateField(auto_now=False, auto_now_add=False, blank=False)
-    image = models.ImageField(upload_to="img/dokter/")
+    image = models.ImageField(upload_to="img/dokter/", default='img/dokter/default.png')
     pengalaman = models.CharField(max_length=3)
     pendidikan = models.CharField(max_length=100)
     rumahsakit = models.ManyToManyField("rumahsakit.RumahSakit", related_name='RumahSakit', blank=False, db_constraint=False) 
@@ -34,9 +34,10 @@ class Doctor(models.Model):
 
 @receiver(models.signals.post_delete, sender=Doctor)
 def auto_delete_image_on_delete(sender, instance, **kwargs):
-    if instance.image:
-        if os.path.isfile(instance.image.path):
-            os.remove(instance.image.path)
+        if instance.image:
+            if os.path.isfile(instance.image.path):
+                if not 'default.png' in instance.image.path:
+                    os.remove(instance.image.path)
 
 @receiver(models.signals.pre_save, sender=Doctor)
 def auto_delete_image_on_change(sender, instance, **kwargs):
